+++ lib/src/utils/format.dart
@@ -2,7 +2,9 @@ import 'dart:io';
 import 'package:dart_style/dart_style.dart';
 import 'package:io/ansi.dart';
 
-final DartFormatter _formatter = DartFormatter();
+final DartFormatter _formatter = DartFormatter(
+  languageVersion: DartFormatter.latestLanguageVersion,
+);
 
 Future<void> formatFile(File? file) async {
   if (file == null) {